+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 #### Bugfixes
   - Windows style line endings are now supported in the config file format (#249)
   - Config files no longer need to end with a line ending (#91)
+  - Removed the last dependency on std::experimental::filesystem (no change in behaviour expected) (#161)
 
 ## v2.23 (2026-04-24)
 #### Features